requirements.txt
##
## Gateway:
##
falcon==3.0.0a2
secp256k1==0.13.2
requests==2.25.1
websocket-client==0.57.0
ecdsa==13.0.0
pycryptodome==3.20.0

##
## Embedded Blockchain libs:
## 
#
# aioxrpy==1.0.0
# eosiopy==0.1.4
# python-bitcoinrpc==1.0
